webSite/backend/requirements.txt
flask
flask-cors
PyYAML